==> 03-image-caption-generator/12-translate-data.ipynb <==
import json
import requests
import time
import pandas as pd

path_dataset = "D:/datasets/flickr-30k-images"
path_captions = "D:/YandexDisk/datasets/flickr-30k-images-metadata.csv"
path_prepared_captions = "D:/flickr-30k-images-metadata.csv"

df = pd.read_csv(path_captions, delimiter='|')
df.head(5)

## Перевод

Квоты — организационные ограничения, которые можно изменить по запросу в техническую поддержку.
Лимиты — технические ограничения, обусловленные особенностями архитектуры Yandex.Cloud. Изменение лимитов невозможно.

Yandex Translate. Квоты (4 апреля 2021):
Вызовов одного метода API в секунду – 20
Символов, отправленных на перевод, в час – 1 млн

IAM_TOKEN = "_____"
FOLDER_ID = "_____"

BEARER = "Bearer " + IAM_TOKEN
HEADERS = {
    "Content-type": "application/json",
    "Authorization": BEARER
}

def translate_with_yandex(text):
    data = {
        "folder_id": FOLDER_ID,
        "texts": text,
        "targetLanguageCode": "ru"
    }

    response = requests.post('https://translate.api.cloud.yandex.net/translate/v2/translate', headers=HEADERS, data=str(data).encode('utf-8'))
    json_text = json.loads(response.text)

    return json_text["translations"][0]["text"]

translate_with_yandex("cloud")

def request_yandex_translate(data, col_name, start_index, end_index):
    start = start_index
    count_iters = 0

    for idx in range(start_index, end_index+1, 20):
        start_time = time.time()

        df_twenty = data[start:idx]
        start = idx

        for idx_twenty in df_twenty.index:
            df_twenty.at[idx_twenty, col_name] = translate_with_yandex(df_twenty.at[idx_twenty, col_name])

        count_iters += 20
        if count_iters > 999:
            data.to_csv("D:/temp.csv", sep='|', encoding='utf-8', index=False)
            count_iters = 0

        time_difference = time.time() - start_time

        if time_difference < 1.0:
            time.sleep(1.01 - time_difference)

## Подсчёт количества знаков

def count_chars_num(data, start_index, end_index, max_count):
    chars_counter = 0
    last_idx = -1

    for idx in range(start_index, end_index+1):
        chars_counter += len(data.iloc[idx, 2])
        if chars_counter > max_count:
            last_idx = idx
            break

    return chars_counter, last_idx

counter, idx = count_chars_num(df, 0, 16460, 1000000)

print(idx)
print(counter)

## Выбрать по миллиону символов

(1) 0 -- 16460
(2) 16461 -- 33234
(3) 33235 -- 50124
(4) 50125 -- 67337
(5) 67338 -- 80713

first_idx = 16461
second_idx = 33234

df_to_trans = df[first_idx:second_idx]
df_to_trans.shape

df_to_trans.to_csv("D:/million-flickr-30k-images-metadata.csv", sep='|', encoding='utf-8', index=False)

## Перевести миллион символов

df_to_trans = pd.read_csv("D:/YandexDisk/datasets/million-flickr-30k-images-metadata.csv.csv", delimiter='|')
df_to_trans.tail(5)

%%time
request_yandex_translate(df_to_trans, " comment", 0, 16772)

df_to_trans.to_csv("D:/russian-flickr-30k-images-metadata-0.csv", sep='|', encoding='utf-8', index=False)

## Объединить переведённые данные

df1 = pd.read_csv("D:/YandexDisk/datasets/russian-flickr-30k-images-metadata-1.csv", delimiter='|')
df2 = pd.read_csv("D:/YandexDisk/datasets/russian-flickr-30k-images-metadata-2.csv", delimiter='|')
df3 = pd.read_csv("D:/YandexDisk/datasets/russian-flickr-30k-images-metadata-3.csv", delimiter='|')
df4 = pd.read_csv("D:/YandexDisk/datasets/russian-flickr-30k-images-metadata-4.csv", delimiter='|')
df5 = pd.read_csv("D:/YandexDisk/datasets/russian-flickr-30k-images-metadata-5.csv", delimiter='|')

df_concated = pd.concat([df1, df2, df3, df4, df5])

df_concated.shape

df_concated.to_csv("D:/russian-flickr-30k-images-metadata.csv", sep='|', encoding='utf-8', index=False)
